code format

+++ App/src/main/java/app/company/CompanyService.java
@@ -1,28 +1,43 @@
 package app.company;
 
 import app.database.Dao;
+import app.ticket.Ticket;
 
 import java.util.List;
 
 public class CompanyService {
 
-    private final Dao<Company> repository;
+    private final Dao<Company> companyRepo;
+    private final Dao<Ticket> ticketsRepo;
 
-    public CompanyService(Dao<Company> repository) {
-        this.repository = repository;
+    public CompanyService(Dao<Company> companyRepo, Dao<Ticket> ticketsRepo) {
+        this.companyRepo = companyRepo;
+        this.ticketsRepo = ticketsRepo;
     }
 
     public List<Company> getCompanies() {
-        return repository.getAll();
+        return companyRepo.getAll();
+    }
+
+    public Company getCompany(int id) {
+        return companyRepo.getById(id);
     }
 
     public boolean createCompany(Company company) {
-        return repository.insert(company);
+        return companyRepo.insert(company);
+    }
+
+    public boolean updateCompany(Company company) {
+        return companyRepo.update(company);
     }
 
     public boolean deleteCompany(int id) {
-        String query = "DELETE FROM companies WHERE id = ?";
-        return repository.delete(query, id);
+        String query = "DELETE FROM tickets WHERE company=?";
+        if (ticketsRepo.delete(query, id)) {
+            query = "DELETE FROM companies WHERE id = ?";
+            return companyRepo.delete(query, id);
+        }
+        return false;
     }
 
 }
